update: Finalizando front end

+++ src/mobile/mark/_components/InfoCard.jsx
@@ -20,6 +20,7 @@ function Card({ data, onRefresh }) {
   const navigate = useNavigate();
 
   const handleDelete = () => {
+    if (!confirm("Deseja excluir essa marca?")) return
     fetch(`${host}/mark/delete/${data.marca}`, {method: "DELETE"})
       .then(() => onRefresh())
       .catch(err => console.error("Houve um erro: " + err))
